Crate wfd

This crate provides safe methods for using Open and Save dialog boxes on Windows.

Structs

DialogParams
The parameters used when displaying a dialog box. All fields are optional and have appropriate default values

OpenDialogResult
The result of an Open Dialog after the user has selected one or more files (or a folder)

SaveDialogResult
The result of a Save Dialog after the user has selected a file

Enums

DialogError
Error returned when showing a dialog fails

Constants

FOS_ALLNONSTORAGEITEMS

FOS_ALLOWMULTISELECT

FOS_CREATEPROMPT

FOS_DEFAULTNOMINIMODE

FOS_DONTADDTORECENT

FOS_FILEMUSTEXIST

FOS_FORCEFILESYSTEM

FOS_FORCEPREVIEWPANEON

FOS_FORCESHOWHIDDEN

FOS_HIDEMRUPLACES

FOS_HIDEPINNEDPLACES

FOS_NOCHANGEDIR

FOS_NODEREFERENCELINKS

FOS_NOREADONLYRETURN

FOS_NOTESTFILECREATE

FOS_NOVALIDATE

FOS_OVERWRITEPROMPT

FOS_PATHMUSTEXIST

FOS_PICKFOLDERS

FOS_SHAREAWARE

FOS_STRICTFILETYPES

FOS_SUPPORTSTREAMABLEITEMS

Functions

open_dialog
Displays an Open Dialog using the provided parameters.

save_dialog
Displays a Save Dialog using the provided parameters.